Version 2.0
I added a user configuration section, as discussed on the vim-dev
mailing list.  With the defaults, most :imap's are not defined, so it
is safer to include this in your plugin directory, if you like using any
of the commands or functions it defines.  Like the rest of the file,
the configuration section is meant as an example to be followed.
The user should copy this section to a separate file, foo.vimrc .

I also added functions to save script-local variables in the script
file, so they are made persistent.